requirements.txt
# Core dependencies
PyQt5==5.15.9
pytesseract==0.3.10
Pillow==10.0.0
pynput==1.7.6
requests==2.31.0
supabase==1.0.3
python-dotenv==1.0.0
pyperclip==1.8.2

# Platform specific
pywin32==306; sys_platform == 'win32'
pyautogui==0.9.54 